app: use ADAravis driver.
`commonDriverMakefile` is used to configure all the dependencies,
including asyn, busy, as well as system libraries (those built in
ADSupport or not). This simplifies greatly src/Makefile. These
dependencies are listed in the RELEASE file. However, no path is given
as it will be determined by the container environment yet to be
configured.

`device.cmd` has been created to be a template for general configuration
of the device, including IOC initialization, ADAravis driver
creation and database loading. This should be reused in all IOCs
start-up scripts created.

`st.cmd` serves as an example of the start-up script that includes
`device.cmd`. It should also include plugin configuration from an
external file. iocInit has been left to be called in it, so that one can
specify whether another script should go before or after initialization.
For instance, some autosave commands must go after.