database: add accelerator standard PV behavior and naming.
Both the behavior and naming come from the previous implementation for
the Basler acA1300-75gm used at the Sirius accelerator. This
implementation was obtained from commit 967176553fe (latest one in any
branch available as of now) of the following repository

https://github.com/lnls-dig/basler-acA1300-75gm-epics-ioc

The basleracA130075gmApp/Db/BaslerSiriusStandard.db was renamed to
`accelerator.db` here. High level aliasing has not been included, since
it shall be verified whether it is really useful in practice.

Some records in the database use `sseq`, which is provided by the EPICS
Calc module. Thus, it has been added as a dependency in the IOC.