autosave: configure autosave for driver and plugins.
`ioc.req` and `plugins.req` have been designed to be configured in the
IOC instance and versioned along with start-up script (beamline or
accelerator configuration repository). This is because they define
whether the camera is using plugins and which plugins they are.
Moreover, `ioc.req` gives the possibility to add more customized
autosave requirement files for a specific setup.