database: add option to limit camera frame rate.
This has been defined through an overload of the `AcquirePeriod` record,
which makes it an pre-initialization setup running before any Channel
Access put command. This setup should also allow us to write-protect the
field for avoiding any attempt to bypass this limit.

The threshold value is defined in the start-up script, as shown in the
`st.cmd` example. If the limit does not need to be applied, you may
either omit the specific variable definition or don't load `limits.cmd`
at all.