autosave: include search from genicam directory.
User defined configuration files currently are supported for local
directory (.) and `basler` subdirectory. However, other GenICam cameras
can be used with this IOC. Therefore, make the `genicam` folder also a
possibility to supersede the previous `basler` directory. The
specification of the `basler` directory is kept for backward
compatibility with existing deploys.