iocsh: load templates with dynamically determined absolute paths.
Templates loaded with relative path requires that we `cd` to the
template directory at the start of the iocsh script. This makes it
harder to define autosave requirement files to be loaded based on the
user's current directory, which is desired when it is mounted as a
volume.

Avoid changing the directory entirely by using absolute path for IOC
configuration files based on `envPaths` definition. This makes it clearer
which loaded files are templates and which are locally defined.

Note that this configuration makes it the entire script independent of
the build and install location of the IOC, since `envPaths` is placed in
a standard directory. It has been chosen to be under `share/misc` as it
is a single application architecture-independent configuration file [1].

[1]: https://refspecs.linuxfoundation.org/FHS_3.0/fhs/ch04s11.html#usrsharemiscMiscellaneousArchitecture