plugins: allow overwriting ArrayData data types.
Data types depend on the pixel depth of the pixel format currently in
use by the detector. This implies the data type we need to provide the
corresponding data. Using a data type larger than needed will increase
network overhead, possibly making it impossible to transmit the desired
traffic.

Make this possible to be overwritten in the main IOC shell script.