iocsh: make IOC binary discoverable in the environment.
This makes possible to remove the hardcoded path used in the shebang,
aiming at removing all the startup iocsh template script dependencies
of the build directory.

A post-install phase is used to create symbolic links instead of
changing the INSTALL_LOCATION. This change is backward compatible, as
the binary can still be found in the previous path.