plugins: define plugin queue sizes by default.
Most camera IOCs will not require a custom queue size for plugins.
Define the previously recommended values as default, that is QSIZE=20
for all plugins but NDPluginHDF5, which takes a 50-element queue. This
makes the template script shorter and easier to maintain.

This change is backward compatible with previous versions, since the
variables are still used in the `plugins.cmd` template.