database: forward rate limit to FrameRate.DRVH.
Both AcquirePeriod and FrameRate can be used to define frame rate. Keep
their limits consistent by forwarding the inverse of period limit to the
FrameRate record.

A better implementation of this feature would to have a FrameRateLimit
record which forwards for both AcquirePeriod's and FrameRate's DRVL and
DRLH fields. Having the limit specified in terms of frame rate would
however would require changing the configuration parameter defined in
the template st.cmd.

To keep backward compatibility, the frame rate is computed based on the
value of the AcquirePeriod.DRVL instead. Changing it at runtime requires
reprocessing FrameRateLimitFwd to forward to FrameRate.DRVH.