plugins: set up pipeline with codec and HDF5
Compression is an important part of the pipeline when cameras are used
to save a long acquisition to disk. Using the HDF5 compression filters
directly is not recommended, since it compresses using a single thread
before writing to disk. Instead, the best approach is to use Codec to
first compress the array, and then forward to HDF5 for direct chunk
write. Configure this processing chain by default.

Because users are used to configuring the HDF5 plugin directly, we
enable NDPluginCodec by default, so it can take arrays and forward them
to HDF5 when it gets enabled. We make the usage of compression through
Codec an opt-in option, in order to have it gradually enabled in
beamlines that have everything ready to consume them.

Compression mode is defined to ensure it is consistent across IOC
restarts. Number of threads is defined as the maximum when compression
is enabled, as we don't have a good estimate of how many threads are
required in general. With multi-threading, it is interesting to ensure
that the output arrays are sorted, as this will affect the order frames
are written to file. When only forwarding arrays, it should be safe to
fallback to a single thread. There is no performance penalty with having
sorting enabled in single-thread mode, as it will first check for out of
order and then sort if needed (which shouldn't happen).